:hammer: REFACTOR: convert project row props to models
