+++ src/pages/Project/Row/models/ProjectRowFileModel.ts
@@ -1,0 +1,8 @@
+import AssetModel from "models/AssetModel";
+import ProjectRowModel from "./ProjectRowModel";
+
+interface ProjectRowFileModel extends ProjectRowModel {
+  asset: AssetModel;
+}
+
+export default ProjectRowFileModel;
