--- src/pages/Project/Row/models/ProjectRowFileProps.ts
@@ -1,8 +1,0 @@
-import AssetModel from "models/AssetModel";
-import ProjectRowProps from "./ProjectRowProps";
-
-interface ProjectRowFileProps extends ProjectRowProps {
-  asset: AssetModel;
-}
-
-export default ProjectRowFileProps;
