+++ src/pages/Project/Row/models/ProjectRowImageModel.ts
@@ -1,0 +1,8 @@
+import AssetModel from "models/AssetModel";
+import ProjectRowModel from "./ProjectRowModel";
+
+interface ProjectRowImageModel extends ProjectRowModel {
+  asset: AssetModel;
+}
+
+export default ProjectRowImageModel;
