--- src/pages/Project/Row/models/ProjectRowImageProps.ts
@@ -1,8 +1,0 @@
-import AssetModel from "models/AssetModel";
-import ProjectRowProps from "./ProjectRowProps";
-
-interface ProjectRowImageProps extends ProjectRowProps {
-  asset: AssetModel;
-}
-
-export default ProjectRowImageProps;
