+++ src/pages/Project/Row/models/ProjectRowModel.ts
@@ -1,6 +1,14 @@
+enum ProjectRowType {
+  Text,
+  Image,
+  Video,
+  File,
+}
+
 interface ProjectRowModel {
   id: string;
   type: number;
 }
 
 export default ProjectRowModel;
+export { ProjectRowType };
