--- src/pages/Project/Row/models/ProjectRowProps.ts
@@ -1,15 +1,0 @@
-enum ProjectRowType {
-  Text,
-  Image,
-  Video,
-  File,
-}
-
-interface ProjectRowProps {
-  id: string;
-  type: ProjectRowType;
-}
-
-export default ProjectRowProps;
-
-export { ProjectRowType };
