+++ src/pages/Project/Row/models/ProjectRowTextModel.ts
@@ -1,0 +1,16 @@
+import ProjectRowModel from "./ProjectRowModel";
+
+enum TextStyle {
+  h1,
+  h2,
+  h3,
+  p,
+}
+
+interface ProjectRowTextModel extends ProjectRowModel {
+  content: string;
+  textStyle: TextStyle;
+}
+
+export default ProjectRowTextModel;
+export { TextStyle };
